adjust clean up & read from absolute file path

+++ build-info-integration-core/src/main/kotlin/org/octopusden/octopus/artifactory/build/info/integration/service/impl/NpmBuildInfoIntegrationServiceImpl.kt
@@ -72,7 +72,7 @@ class NpmBuildInfoIntegrationServiceImpl(
 
         buildInfoService.uploadBuildInfo(mergedBuildInfo)
 
-        if (buildInfoConfig.cleanupNpmBuildInfo) {
+        if (buildInfoConfig.cleanupNpmBuildInfo && npmBuildInfo != null) {
             buildInfoService.deleteBuildInfo(buildInfoConfig.npmBuildName, listOf(buildInfoConfig.buildNumber))
         }
